Refactored the LogRow into its own view for reuse.
